requirements.txt
imbalanced_learn==0.12.0
imblearn==0.0
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
plotly==5.18.0
scikit_learn==1.4.1.post1
seaborn==0.13.2
streamlit==1.31.0
